+++ fractal_server/app/routes/api/v1/project.py
@@ -42,6 +42,7 @@ from ._aux_functions import _get_dataset_check_owner
 from ._aux_functions import _get_project_check_owner
 from ._aux_functions import _get_submitted_jobs_statement
 from ._aux_functions import _get_workflow_check_owner
+from ._aux_functions import _raise_if_v1_is_read_only
 from ._aux_functions import clean_app_job_list_v1
 
 router = APIRouter()
@@ -80,7 +81,7 @@ async def create_project(
     """
     Create new poject
     """
-
+    _raise_if_v1_is_read_only()
     # Check that there is no project with the same user and name
     await _check_project_exists(
         project_name=project.name, user_id=user.id, db=db
@@ -120,6 +121,7 @@ async def update_project(
     user: User = Depends(current_active_user),
     db: AsyncSession = Depends(get_async_db),
 ):
+    _raise_if_v1_is_read_only()
     project = await _get_project_check_owner(
         project_id=project_id, user_id=user.id, db=db
     )
@@ -148,6 +150,7 @@ async def delete_project(
     """
     Delete project
     """
+    _raise_if_v1_is_read_only()
     project = await _get_project_check_owner(
         project_id=project_id, user_id=user.id, db=db
     )
@@ -249,16 +252,8 @@ async def apply_workflow(
     user: User = Depends(current_active_verified_user),
     db: AsyncSession = Depends(get_async_db),
 ) -> Optional[ApplyWorkflowReadV1]:
-
+    _raise_if_v1_is_read_only()
     settings = Inject(get_settings)
-    if settings.FRACTAL_API_V1_MODE == "include_without_submission":
-        raise HTTPException(
-            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
-            detail=(
-                "Legacy API is still accessible, "
-                "but the submission of legacy jobs is not available."
-            ),
-        )
 
     # Remove non-submitted V1 jobs from the app state when the list grows
     # beyond a threshold